+++ packages/tsurlfilter/src/rules/declarative-converter/filter-converter.ts
@@ -98,6 +98,7 @@
 
 import { getErrorMessage } from '../../common/error';
 import type { NetworkRule } from '../network-rule';
+import { isSafeRule } from '../declarative-converter-utils/is-safe-dnr-rule';
 
 import {
     type IRuleSet,
@@ -416,10 +417,7 @@ export class DeclarativeFilterConverter implements IFilterConverter {
 
         const rulesHashMap = new RulesHashMap(listOfRulesWithHash);
 
-        // calculate number of unsafe rules only for dynamic rules
-        const unsafeRulesCount = DeclarativeFilterConverter.COMBINED_RULESET_ID === ruleSetId
-            ? declarativeRules.filter((r) => !DeclarativeRulesConverter.isSafeRule(r)).length
-            : 0;
+        const unsafeRulesCount = declarativeRules.filter((r) => !isSafeRule(r)).length;
 
         const regexRulesCount = declarativeRules.filter((r) => DeclarativeRulesConverter.isRegexRule(r)).length;