+++ packages/tswebextension/src/lib/mv3/background/web-request-api.ts
@@ -650,7 +650,23 @@ export class WebRequestApi {
             return;
         }
 
-        const rule = matchingResult.getBasicResult();
+        /**
+         * When handling document blocking, we need to retrieve the rule responsible for the blocking.
+         *
+         * MatchingResult.getBasicResult() is not good enough, because it may return an exception rule,
+         * while a blocking rule is still applied in MV3. For example, if there are two rules:
+         * `||example.com^$document` and
+         * `@@||example.com^$generichide`,
+         * the first rule is converted and applied as DNR rule by browser (so here we are handling the error),
+         * the second one is not related to the blocking and error handling.
+         *
+         * This behaviour is expected, correct, and consistent with the CoreLibs apps:
+         * https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3260.
+         *
+         * TODO: fix incorrect logic for the MV2 extension:
+         * https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3262.
+         */
+        const rule = matchingResult.getDocumentBlockingResult();
         if (!rule) {
             return;
         }